// fornt-end/types/enums.ts
export enum StatusLivro {
  DISPONIVEL = 'DISPONIVEL',
  EMPRESTADO = 'EMPRESTADO',
  RESERVADO = 'RESERVADO',
}

export enum Etiqueta {
  BRANCO = 'BRANCO',
  AMARELO = 'AMARELO',
  VERMELHO = 'VERMELHO',
}

export enum Role {
  ADMIN = 'ADMIN',
  BIBLIOTECARIO = 'BIBLIOTECARIO',
  MEMBRO = 'MEMBRO',
}

export enum TipoMembro {
  ESTUDANTE = 'ESTUDANTE',
  PROFESSOR = 'PROFESSOR',
}